pyproject.toml
[tool.poetry]
name = "talentdb"
version = "0.1.0"
description = ""
authors = ["k1m0ch1 <yahya.kimochi@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.32.3"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
